Update routing and UI components for contests and profile pages; add CommonAppBar widget for consistent app bar design. Remove deprecated contests page and refactor side menu navigation to use GoRouter. Enhance profile page with new layout and problem data structure.
