+++ a2sv_hub_mobile/.gitignore
@@ -41,3 +41,6 @@ app.*.map.json
 /android/app/debug
 /android/app/profile
 /android/app/release
+
+CXX build folder
+/android/app/.cxx/
