+++ a2sv_hub_mobile/lib/core/presentation/routes/router.dart
@@ -1,7 +1,7 @@
 import 'package:a2sv_hub/features/auth/presentation/pages/login_page.dart';
 import 'package:a2sv_hub/features/auth/presentation/pages/role_page.dart';
 import 'package:a2sv_hub/features/auth/presentation/pages/signup_page.dart';
-import 'package:a2sv_hub/features/contests/presentation/pages/contests.dart';
+import 'package:a2sv_hub/features/contests/presentation/pages/contest_page.dart';
 import 'package:a2sv_hub/features/contests/presentation/pages/profile.dart';
 import 'package:go_router/go_router.dart';
 
@@ -10,7 +10,7 @@ import '../../../features/presentation/pages/problem_page.dart';
 import '../../../features/presentation/pages/profile_page.dart';
 import 'app_routes.dart';
 
-final router = GoRouter(initialLocation: '/profile', routes: <RouteBase>[
+final router = GoRouter(initialLocation: '/role', routes: <RouteBase>[
   //! Auth -------------------------------------------------------------------
   GoRoute(
     path: Routes.role,
@@ -27,6 +27,9 @@ final router = GoRouter(initialLocation: '/profile', routes: <RouteBase>[
   GoRoute(
     path: Routes.contests,
     builder: (context, state) => const ContestPage(),
+   
+  ),
+  GoRoute(
     path: Routes.home,
     builder: (context, state) => const HomePage(),
   ),
@@ -36,6 +39,6 @@ final router = GoRouter(initialLocation: '/profile', routes: <RouteBase>[
   ),
   GoRoute(
     path: Routes.profile,
-    builder: (context, state) => const ProfilePage(),
+    builder: (context, state) =>  ProfilePage(),
   ),
 ]);
