+++ a2sv_hub_mobile/lib/core/presentation/widgets/common_app_bar.dart
@@ -1,0 +1,104 @@
+import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+import '../routes/app_routes.dart';
+
+class CommonAppBar extends StatelessWidget implements PreferredSizeWidget {
+  final String? title;
+  final List<Widget>? additionalActions;
+  final VoidCallback? onMenuPressed;
+  final bool showBackButton;
+
+  const CommonAppBar({
+    Key? key,
+    this.title,
+    this.additionalActions,
+    this.onMenuPressed,
+    this.showBackButton = false,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return AppBar(
+      backgroundColor: Colors.transparent,
+      elevation: 0,
+      centerTitle: false,
+      leading: showBackButton
+          ? IconButton(
+              icon: const Icon(Icons.arrow_back, color: Colors.black87),
+              onPressed: () => Navigator.of(context).pop(),
+            )
+          : Builder(builder: (BuildContext context) {
+              return IconButton(
+                icon: const Icon(Icons.menu, color: Colors.black87),
+                onPressed: onMenuPressed ??
+                    () {
+                      Scaffold.of(context).openDrawer();
+                    },
+              );
+            }),
+      title: title != null
+          ? Text(
+              title!,
+              style: const TextStyle(
+                color: Colors.black,
+                fontSize: 24,
+                fontWeight: FontWeight.bold,
+              ),
+            )
+          : null,
+      actions: [
+        const SizedBox(width: 50),
+        IconButton(
+          icon: const Icon(Icons.search, color: Colors.black87),
+          onPressed: () {},
+        ),
+        const Spacer(),
+        Stack(
+          alignment: Alignment.center,
+          children: [
+            IconButton(
+              icon: const Icon(Icons.notifications_outlined,
+                  color: Colors.black87),
+              onPressed: () {},
+            ),
+            Positioned(
+              top: 8,
+              right: 8,
+              child: Container(
+                width: 18,
+                height: 18,
+                decoration: BoxDecoration(
+                  color: Colors.red,
+                  borderRadius: BorderRadius.circular(9),
+                ),
+                child: const Center(
+                  child: Text(
+                    '3',
+                    style: TextStyle(
+                      color: Colors.white,
+                      fontSize: 10,
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+                ),
+              ),
+            ),
+          ],
+        ),
+        const SizedBox(width: 8),
+        GestureDetector(
+          onTap: () => context.go(Routes.profile),
+          child: const CircleAvatar(
+            radius: 16,
+            backgroundImage: AssetImage('assets/images/profile.png'),
+          ),
+        ),
+        const SizedBox(width: 16),
+        if (additionalActions != null) ...additionalActions!,
+      ],
+    );
+  }
+
+  @override
+  Size get preferredSize => const Size.fromHeight(kToolbarHeight);
+}
