+++ a2sv_hub_mobile/lib/features/presentation/pages/profile_page.dart
@@ -1,12 +1,38 @@
+import 'package:a2sv_hub/core/presentation/widgets/common_app_bar.dart';
+import 'package:a2sv_hub/core/presentation/widgets/sidemenu.dart';
+import 'package:a2sv_hub/features/contests/presentation/widgets/contest_card.dart';
+import 'package:a2sv_hub/features/contests/presentation/widgets/contest_progress_card.dart';
+import 'package:a2sv_hub/features/contests/presentation/widgets/header.dart';
+import 'package:a2sv_hub/features/contests/presentation/widgets/notification_modal.dart';
+import 'package:a2sv_hub/features/contests/presentation/widgets/profile_modal.dart';
+import 'package:a2sv_hub/features/contests/presentation/widgets/user_info_card.dart';
 import 'package:a2sv_hub/features/presentation/widgets/consistency_grid.dart';
 import 'package:flutter/material.dart';
 import '../widgets/activity_grid.dart';
 import '../widgets/contest_item.dart';
 import '../widgets/contest_progress_chart.dart';
+import '../widgets/problem_card.dart';
 import '../widgets/profile_header.dart';
 import '../widgets/profile_stat.dart';
 import '../widgets/profile_tab_bar.dart';
 import '../widgets/solved_problem_item.dart';
+
+// Model class for problems
+class Problem {
+  final String id;
+  final String title;
+  final String category;
+  final String difficulty;
+  final bool isCompleted;
+
+  Problem({
+    required this.id,
+    required this.title,
+    required this.category,
+    required this.difficulty,
+    this.isCompleted = false,
+  });
+}
 
 class ProfilePage extends StatefulWidget {
   const ProfilePage({Key? key}) : super(key: key);
@@ -17,6 +43,23 @@ class ProfilePage extends StatefulWidget {
 
 class _ProfilePageState extends State<ProfilePage> {
   int _selectedTabIndex = 0;
+
+  bool showNotifications = false;
+  bool showProfileMenu = false;
+
+  void toggleNotifications() {
+    setState(() {
+      showNotifications = !showNotifications;
+      showProfileMenu = false;
+    });
+  }
+
+  void toggleProfileMenu() {
+    setState(() {
+      showProfileMenu = !showProfileMenu;
+      showNotifications = false;
+    });
+  }
 
   // Example data
   final Map<String, List<bool>> _activityData = {
@@ -54,101 +97,39 @@ class _ProfilePageState extends State<ProfilePage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Colors.white,
-      appBar: AppBar(
-        backgroundColor: Colors.transparent,
-        elevation: 0,
-        leading: IconButton(
-          icon: const Icon(Icons.menu, color: Colors.black87),
-          onPressed: () {},
-        ),
-        actions: [
-          const SizedBox(width: 50),
-          IconButton(
-            icon: const Icon(Icons.search, color: Colors.black87),
-            onPressed: () {},
-          ),
-          const Spacer(),
-          Stack(
-            alignment: Alignment.center,
-            children: [
-              IconButton(
-                icon: const Icon(Icons.notifications_outlined,
-                    color: Colors.black87),
-                onPressed: () {},
-              ),
-              Positioned(
-                top: 8,
-                right: 8,
-                child: Container(
-                  width: 18,
-                  height: 18,
-                  decoration: BoxDecoration(
-                    color: Colors.red,
-                    borderRadius: BorderRadius.circular(9),
-                  ),
-                  child: const Center(
-                    child: Text(
-                      '3',
-                      style: TextStyle(
-                        color: Colors.white,
-                        fontSize: 10,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(width: 8),
-          const CircleAvatar(
-            radius: 16,
-            backgroundImage: AssetImage('assets/images/profile.png'),
-          ),
-          const SizedBox(width: 16),
-        ],
+      drawer: const Drawer(
+        backgroundColor: Colors.white,
+        child: SideMenu(),
       ),
-      body: Padding(
-        padding: const EdgeInsets.all(8.0),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            const SizedBox(height: 16),
-            Text(
-              'Profile',
-              style: TextStyle(
-                fontSize: 24,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
+      backgroundColor: Colors.white,
+      appBar: const CommonAppBar(title: 'Profile'),
+      body: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          // Profile header with user info
+          const ProfileHeader(
+            name: 'Amy Santiago Watson',
+            role: 'Student',
+            roleTag: 'A2SV',
+            avatarUrl: 'assets/images/profile.png',
+          ),
 
-            const SizedBox(height: 16),
-            // Profile header with user info
-            const ProfileHeader(
-              name: 'Amy Santiago Watson',
-              role: 'Student',
-              roleTag: 'A2SV',
-              avatarUrl: 'assets/images/profile.png',
-            ),
+          // Tab navigation
+          ProfileTabBar(
+            tabs: const ['Profile', 'Problems', 'Contests'],
+            selectedIndex: _selectedTabIndex,
+            onTabSelected: (index) {
+              setState(() {
+                _selectedTabIndex = index;
+              });
+            },
+          ),
 
-            // Tab navigation
-            ProfileTabBar(
-              tabs: const ['Profile', 'Problems', 'Contests'],
-              selectedIndex: _selectedTabIndex,
-              onTabSelected: (index) {
-                setState(() {
-                  _selectedTabIndex = index;
-                });
-              },
-            ),
-
-            // Tab content
-            Expanded(
-              child: _buildTabContent(),
-            ),
-          ],
-        ),
+          // Tab content
+          Expanded(
+            child: _buildTabContent(),
+          ),
+        ],
       ),
     );
   }
@@ -173,7 +154,19 @@ class _ProfilePageState extends State<ProfilePage> {
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
           // Stats row
-        
+          Row(
+            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+            children: const [
+              ProfileStat(
+                label: 'Problems',
+                value: '1,299',
+              ),
+              ProfileStat(
+                label: 'Attendance',
+                value: '312',
+              ),
+            ],
+          ),
 
           const SizedBox(height: 24.0),
 
@@ -260,105 +253,118 @@ class _ProfilePageState extends State<ProfilePage> {
   }
 
   Widget _buildProblemsTab() {
-    return ListView(
-      children: [
-        // Problem items
-        SolvedProblemItem(
-          title: 'F - Ilya and Queries',
-          category: 'Contest',
-          difficulty: 'Hard',
-          isCompleted: true,
+    final List<Problem> problems = [
+      Problem(
+        id: '1',
+        title: 'F - Ilya and Queries',
+        category: 'Contest',
+        difficulty: 'Hard',
+        isCompleted: true,
+      ),
+      Problem(
+        id: '2',
+        title: '2115. Find all possible recipes from given suppliers',
+        category: 'Topological Sort',
+        difficulty: 'Medium',
+        isCompleted: true,
+      ),
+      Problem(
+        id: '3',
+        title: 'Valid Parenthesis',
+        category: 'Stack',
+        difficulty: 'Easy',
+        isCompleted: true,
+      ),
+      Problem(
+        id: '4',
+        title: 'Merge Two Sorted Lists',
+        category: 'Linked List',
+        difficulty: 'Easy',
+        isCompleted: false,
+      ),
+      Problem(
+        id: '5',
+        title: 'Minimum Path Sum',
+        category: 'Dynamic Programming',
+        difficulty: 'Medium',
+        isCompleted: false,
+      ),
+    ];
+
+    return ListView.builder(
+      padding: const EdgeInsets.all(16.0),
+      itemCount: problems.length,
+      itemBuilder: (context, index) {
+        final problem = problems[index];
+        return ProblemCard(
+          title: problem.title,
+          category: problem.category,
+          difficulty: problem.difficulty,
+          isCompleted: problem.isCompleted,
           onTap: () {},
-          onEditTap: () {},
-        ),
-        SolvedProblemItem(
-          title: '2115. Find all possible recipes from given suppliers',
-          category: 'Topological Sort ...',
-          difficulty: 'Medium',
-          isCompleted: true,
-          onTap: () {},
-          onEditTap: () {},
-        ),
-        SolvedProblemItem(
-          title: 'Valid Parenthesis',
-          category: 'Stack',
-          difficulty: 'Easy',
-          isCompleted: true,
-          onTap: () {},
-          onEditTap: () {},
-        ),
-        SolvedProblemItem(
-          title: 'Merge Two Sorted Lists',
-          category: 'Linked List',
-          difficulty: 'Easy',
-          isCompleted: false,
-          onTap: () {},
-          onEditTap: () {},
-        ),
-        SolvedProblemItem(
-          title: 'Minimum Path Sum',
-          category: 'Dynamic Programming',
-          difficulty: 'Medium',
-          isCompleted: false,
-          onTap: () {},
-          onEditTap: () {},
-        ),
-      ],
+          onToggleComplete: () {},
+          onOpenDetails: () {},
+        );
+      },
     );
   }
 
   Widget _buildContestsTab() {
-    return SingleChildScrollView(
-      padding: const EdgeInsets.all(16.0),
-      child: Column(
-        children: [
-          // Contest Progress Chart
-          ContestProgressChart(
-            title: 'Contest Progress',
-            subtitle: 'Coder #481',
-            data: _contestProgressData,
+    return Stack(
+      children: [
+        SingleChildScrollView(
+          child: Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 18.0),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                // ContestHeader(
+                //   onNotificationTap: toggleNotifications,
+                //   onProfileTap: toggleProfileMenu,
+                // ),
+                // const UserInfoCard(),
+                const ContestProgressCard(),
+                const SizedBox(height: 29),
+                SizedBox(
+                  height: 500, // Constrain the height of the ListView
+                  child: ListView.separated(
+                    physics: const NeverScrollableScrollPhysics(),
+                    itemCount: 5,
+                    separatorBuilder: (context, index) =>
+                        const SizedBox(height: 30),
+                    itemBuilder: (context, index) {
+                      return SizedBox(
+                        height: 139,
+                        child: ContestCard(
+                          title: index % 2 == 0
+                              ? '133. A2SV Ghana G6 - Round #5'
+                              : '132. A2SV Remote Contest #8',
+                          subtitle: index % 2 == 0
+                              ? '6 problems . 14h ago'
+                              : '6 problems . 14h ago',
+                        ),
+                      );
+                    },
+                  ),
+                ),
+              ],
+            ),
           ),
-
-          const SizedBox(height: 24.0),
-
-          // Contest List
-          ContestItem(
-            title: '133. A2SV Ghana G6 - Round #5',
-            date: '14h ago',
-            problems: 6,
-            onTap: () {},
-            onDismiss: () {},
-          ),
-          ContestItem(
-            title: '132. A2SV Remote Contest #8',
-            date: '14h ago',
-            problems: 6,
-            onTap: () {},
-            onDismiss: () {},
-          ),
-          ContestItem(
-            title: '131. A2SV Ghana G6 - Round #4',
-            date: '70h ago',
-            problems: 6,
-            onTap: () {},
-            onDismiss: () {},
-          ),
-          ContestItem(
-            title: '130. A2SV Ghana G6 - Round #3',
-            date: '14h ago',
-            problems: 6,
-            onTap: () {},
-            onDismiss: () {},
-          ),
-          ContestItem(
-            title: '129. A2SV Remote Contest #7',
-            date: '36h ago',
-            problems: 6,
-            onTap: () {},
-            onDismiss: () {},
-          ),
-        ],
-      ),
+        ),
+         Align(
+              alignment: Alignment.centerRight,
+              child: Padding(
+                padding: const EdgeInsets.only(right: 12.0),
+                child: Image.asset(
+                  'assets/images/theme.jpg',
+                  width: 50,
+                  height: 50,
+                ),
+              ),
+            ),
+            if (showNotifications) const NotificationModal(),
+            if (showProfileMenu) const ProfileModal(),
+      ],
     );
   }
