+++ a2sv_hub_mobile/lib/main.dart
@@ -34,7 +34,7 @@ class MyApp extends StatelessWidget {
           colorScheme: ColorScheme.fromSeed(seedColor: const Color(0xFF2065D1)),
           useMaterial3: true,
         ),
-      ],
+      ),
     );
   }
 }
